scripts/openstack-cleanup/requirements.txt
openstacksdk>=0.43.0
